frontend/pyproject.toml
[project]
name = "project_hermes_frontend"
version = "0.1.0"
description = "Frontend for Project Hermes travel planning system"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "streamlit>=1.24.0",
    "requests>=2.31.0",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
